[代码质量与功能优化]: 主要代码质量优化、修复潜在漏洞及功能改进
- 更新 `.clang-tidy` 配置文件，调整代码检查规则，禁用部分不适用的检查项，增加更多潜在问题检测规则，提高代码静态分析覆盖率
- 优化 GitHub Actions 工作流，扩展忽略规则，避免误报，提升 CI 检测效率
- 改进多种搜索算法实现（顺序、二分、斐波那契等），统一变量命名规范，增强代码可读性
- 调整 Curl 相关模块（文件工具、HTTP 客户端等），优化类型使用（如将 `double` 改为 `int64_t`），增加参数校验，提升代码健壮性
- 完善设计模式示例（工厂模式、单例模式等），添加单元测试，增强线程安全性验证
- 优化服务器模块（套接字操作等），改进函数参数命名，规范代码风格
- 重构线程相关模块（线程类、线程池等），调整类型定义（如使用 `uint64_t` 等），增加 `[[nodiscard]]` 注解，提升代码语义清晰度
- 更新工具类和宏定义（如 `ScopeGuard`、对象操作宏等），增强资源管理安全性，规范拷贝/移动语义
- 新增 `utils/utils.hpp` 文件，定义 `UNUSED` 宏，便于处理未使用变量警告
- 修正 README 文档中部分内容顺序和格式，提升文档可读性，确保信息准确完整

+++ Curl/file_utils.cc
@@ -40,7 +40,7 @@ void assertFileData(const std::string &filename, const std::string &data)
     EXPECT_EQ(fileData, data);
 }
 
-auto formatBytes(double value, int precision) -> std::string
+auto formatBytes(int64_t value, int precision) -> std::string
 {
     std::vector<std::string> units = {"B", "KB", "MB", "GB", "TB", "PB", "EB", "ZB", "YB"};
