+++ Curl/file_utils.hpp
@@ -8,4 +8,4 @@ void removeFile(const std::string &filename);
 
 void assertFileData(const std::string &filename, const std::string &data);
 
-auto formatBytes(double value, int precision = 2) -> std::string;
+auto formatBytes(int64_t value, int precision = 2) -> std::string;
